Always create clip nodes for iframes
When an iframe pipeline is not ready, we previously didn't create the
clip node or process the pipeline for that iframe. This results in the
ClipScrollTree having "empty" ClipNodes. These are nodes for which we
have allocated an index, but never process its display list item. We
avoid this by always adding the ClipNodes for these iframes.

Fixes #2898.